Move endpoint to own class, now works as true middleware based on external flask app
